% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{combine}
\alias{combine}
\title{Combines two layout matricies on the matching the pattern value.}
\usage{
combine(M1, M2, pattern)
}
\arguments{
\item{M1}{The target layout-matrix containing a string matching the pattern
specified.}

\item{M2}{The layout-matrix to be inserted in the target layout-matrix.}

\item{pattern}{A string present in the layout-matrix M1, that specifies where
the layout-matrix M2 will be inserted.}
}
\description{
Combines two layout matricies on the matching the pattern value.
}
\examples{
# Create a layout matrix
M1 <- LM( X=c(1,"A",2,"A") , C=2 , H=c(1,2) )
M2 <- LM( X=1:6 , C=2 , W=c(3,4) , H=c(3,2,6) )
vis(M1)
vis(M2)

# Combine the two layout matrix on index 'A'
M3 <- combine( M1 = M1, M2 = M2, pattern = "A")
vis(M3)
}
