% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{mypar}
\alias{mypar}
\title{AAA}
\usage{
mypar(
  R = 1,
  C = 1,
  cex.lab = 1,
  cex.axis = 1,
  cex.main = 1,
  font.main = 1,
  mar = c(2.5, 2.5, 1.6, 1.1),
  mgp = c(1.1, 0.3, 0),
  tcl = -0.2,
  las = 1,
  ...
)
}
\description{
AAA
}
\details{
Sets better default values for the base R grafics

AAA
}
