% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{new_layout}
\alias{new_layout}
\title{Creates a new layout to display plots.}
\usage{
new_layout(X)
}
\arguments{
\item{X}{A layout-matrix created with the function \code{LM}.}
}
\description{
Creates a new layout to display plots.
}
