% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{select}
\alias{select}
\title{Subsets row, columns and cells matching desired filters. The extracted matrix
is expected to be a rectangular matrix. single-row/column matrices are allowed.}
\usage{
select(X, rows = NULL, cols = NULL, pattern = NULL, inverse = FALSE)
}
\arguments{
\item{X}{A layout-matrix object created with \code{LM()} function.}

\item{rows}{A vector containing the row indexe(s) to extract from the layout-
matrix \code{X}. This parameter #' can be used in conjuction with \code{pattern} and
\code{cols} for more flexible subsetting.}

\item{cols}{A vector containing the column indexe(s) to extract from the
layout-matrix \code{X}. This parameter can be used in conjuction with \code{rows} and
\code{pattern} for more flexible subsetting.}

\item{pattern}{A pattern to extract from the layout-matrix \code{X}. This parameter
can be used in conjuction with \code{rows} and \code{cols} for more flexible subsetting.}
}
\value{
This function returns a layout matrix containing the selected pattern,
row and/or column.
}
\description{
Subsets row, columns and cells matching desired filters. The extracted matrix
is expected to be a rectangular matrix. single-row/column matrices are allowed.
}
\examples{
# Create a layout matrix
C <- LM( X=1:6 , C=2 , W=c(3,4) , H=c(3,2,7) )

# Subset layout matrix with cells containing the value "2"
select(X = C, pattern = "2" )

# Subset layout matrix with cells containing the value "2" or "4"
select(X = C, pattern = c("2","4") )

# Subset layout matrix to contain the 2nd row
select(X = C, rows = 2:3 )

# Subset layout matrix to contain the 1st column
select(X = C, cols = 1 )

#' # Subset layout matrix to contain the 1st column AND 2nd row
select(X = C, cols = 1 , rows = 2)
}
